Core support for double values in colHeight and colRadius. isRooted() replaced by isMovementDisabled() in AI.
git-svn-id: http://svn.l2jserver.com/trunk@3832 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/com/l2jserver/gameserver/network/serverpackets/ServerObjectInfo.java
@@ -28,7 +28,7 @@ public final class ServerObjectInfo extends L2GameServerPacket
 	private int _x, _y, _z, _heading;
 	private int _idTemplate;
 	private boolean _isAttackable;
-	private int _collisionHeight, _collisionRadius;
+	private double _collisionHeight, _collisionRadius;
 	private String _name;
 
 	public ServerObjectInfo(L2Npc activeChar, L2Character actor)
